refactor(mapper): ReservationTime에 관한 dto <-> domain 변환 로직을 책임지는 mapper 구현
